requirements.txt
fastapi
uvicorn
transformers
langchain
requests
mysql-connector-python
python-dateutil
sqlalchemy
